Database hooked up
connection.php was added in the folder 'includes' so it can be included in everything.
The messages you send on the index.php page will now be saved in the database.
There is a label for username, this feature will be added.
The user also has to see what's inside the database. This needs to be added aswell.